Solve weirdness with restarting auton in sim.
When we start auton, we reset our odometry to a predefined position--
under the presumption that our humans have placed us at this well-known
position prior to starting a match.

So both our DrivetrainSim and odometry should agree-- we are at our
predefined starting position, with 0 velocity or previous state.

It kinda seems like the only reliable way to reset the DrivetrainSim is
to re-instantiate it? I guess? Or maybe I was chasing my tail earlier, I
definitely was accidentally not retaining the "reset" callback object
thingy and it was getting garbage collected, resulting in a missing
reset callback on my encoders...

In any case, I can now reliably see the odometry reset on auton start,
without any weird artifacting like I was seeing before, and this commit
message is getting absurdly long, so I'ma leave it here. Deuces.