"Long shot" autonomous
Was tinkering to see if I could pick up all 3 of our balls from our side
of the field and deliver them all before auton is over. If we drive like
hell, and the pickup and shooter both work 100% and fast, then yes.

I kinda think not, but hey, it's a nice dream right?